requirements.txt
chromadb
google-generativeai
python-dotenv
